refactor(simulator): split into parts/modules
