+++ simulator/src/lib.rs
@@ -1,1 +1,5 @@
+pub mod fake_event;
+pub mod fake_events;
+pub mod fake_peer;
+pub mod fake_peer_id;
 pub mod one_real_node_cluster;
